+++ lib/teamlab/modules/mail/mail_helpcenter.rb
@@ -1,0 +1,8 @@
+module Teamlab
+  # Methods for working with mail helpcenter
+  module MailHelpCenter
+    def get_html_of_help_center
+      @request.get(%w[helpcenter])
+    end
+  end
+end