+++ spec/lib/mail/mail_helpcenter_spec.rb
@@ -1,0 +1,11 @@
+require 'spec_helper'
+
+describe '[Mail] HelpCenter' do
+  let(:teamlab_module) { :mail }
+
+  describe '#get_html_of_help_center' do
+    it_should_behave_like 'an api request' do
+      let(:command) { :get_html_of_help_center }
+    end
+  end
+end